+++ botcli/botcli.go
@@ -100,3 +100,17 @@ func (self *BotCli) AddCommand(cmd *cobra.Command, callback Callback) {
 	self.RootCmd.AddCommand(cmd)
 	self.cmdsMap[cmd.Use] = callback
 }
+
+// Store a custom program setting in the given bot. The setting is specific to your application.
+//
+// The setting is stored using Bot.SetUiConfig() and the key is prefixed with BotCli.AppName.
+func (self *BotCli) SetConfig(bot *deltachat.Bot, key, value string) {
+	bot.SetUiConfig(self.AppName+"."+key, value)
+}
+
+// Get a custom program setting from the given bot. The setting is specific to your application.
+//
+// The setting is retrieved using Bot.GetUiConfig() and the key is prefixed with BotCli.AppName.
+func (self *BotCli) GetConfig(bot *deltachat.Bot, key string) string {
+	return bot.GetUiConfig(self.AppName + "." + key)
+}